backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI hello world service"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.17.0",
    "asyncpg>=0.30.0",
    "fastapi[standard]",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.1.1",
    "sqlalchemy[asyncio]>=2.0.44",
    "uvicorn[standard]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.hatch.build.targets.sdist]
include = ["app", "README.md", "pyproject.toml"]

